added Ezreal class to handle discord embeds.
